requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openpyxl==3.1.2
pdfplumber==0.11.0
python-multipart==0.0.9
python-docx==1.1.2
pypdf==4.3.1
